:recycle: REFACTOR: splitting functions output of single file based on responsbility
